Chỉnh sửa load cart, thêm giỏ hàng, thay đổi interface discount

+++ src/app/user/components/user-home/user-cart/user-cart.component.ts
@@ -8,6 +8,8 @@ import { DeliveryOrderService } from 'src/app/services/delivery-order.service';
 import { OrderService } from 'src/app/services/order.service';
 import { Product } from 'src/app/interfaces/product';
 import { ToastService } from 'src/app/services/toast.service';
+import Swal from 'sweetalert2';
+import { DiscountService } from 'src/app/services/discount.service';
 
 @Component({
   selector: 'app-user-cart',
@@ -37,15 +39,9 @@ export class UserCartComponent implements OnInit {
     private orderService : OrderService,
     private cartDetailService : CartDetailService,
     private toastService: ToastService,
+    private discountService : DiscountService,
   ) {
-    this.deliveryOrderService.userCart$.subscribe(cart => {
-      this.cartByUser = cart;
-      console.log(this.cartByUser);
-      // Xử lý khi có sự thay đổi trong giỏ hàng người dùng
-    });
-    this.orderService.userOrderCache$.subscribe(order => {
-      this.orderByUser = order;
-    });
+    
     const productsString = localStorage.getItem('products');
     if (productsString) {
       const parsedProducts: Product[] = JSON.parse(productsString);
@@ -53,12 +49,36 @@ export class UserCartComponent implements OnInit {
     } else {
       console.error('No products found in local storage or the value is null.');
     }
+    this.deliveryOrderService.userCart$.subscribe(cart => {
+      this.cartByUser = cart;
+    });
+    this.orderService.userOrderCache$.subscribe(order => {
+      this.orderByUser = order;
+    });
     this.cartDetailService.orderDetails$.subscribe(orderDetails => {
       this.orderDetailByUser = orderDetails;
+      if(orderDetails){
+        this.setCartInfomationArrays();
+      }
     });
     
     
   }
+
+  ngOnInit(): void {
+    this.cartService.getProvince();
+    this.cartService.provinceData$.subscribe(data => {
+      this.Provinces = Object.values(data);
+    });
+    this.cartService.districtData$.subscribe(data => {
+      this.Districts = Object.values(data);
+    });
+    this.cartService.wardData$.subscribe(data => {
+      this.Wards = Object.values(data);
+    });
+   
+  }
+
   @ViewChild('likeBtn') likeBtn!: ElementRef;
   @ViewChild('minusBtn') minusBtn!: ElementRef;
   @ViewChild('plusBtn') plusBtn!: ElementRef;
@@ -70,6 +90,7 @@ export class UserCartComponent implements OnInit {
   }
 
   setCartInfomationArrays(){
+    this.cartInformationArray = [];
     if (this.orderDetailByUser && this.products) {
       for (let i = 0; i < this.orderDetailByUser.length; i++) {
         const orderDetail = this.orderDetailByUser[i];
@@ -170,19 +191,28 @@ export class UserCartComponent implements OnInit {
 
   deleteCartDetail(cart : any) : void{
     const orderDetailId = cart.orderDetailId;
-    this.cartDetailService.delete(orderDetailId).subscribe({
-      next: (response) => {
-        if (response.message === "Delete is success") {
-          this.toastService.showTimedAlert('Xóa thành công', '', 'success', 2000);
+    this.toastService.showConfirmAlert('Bạn chắc chắn xóa?', '', 'warning')
+    .then((result) => {
+      if (result.isConfirmed) {
+        this.cartDetailService.delete(orderDetailId).subscribe({
+          next: (response) => {
+            if (response.message === "Delete is success") {
+              this.cartDetailService.getUserOrderDetail(this.orderByUser?.orderId as number).subscribe();
+              this.toastService.showTimedAlert('Xóa thành công', '', 'success', 2000);
+            }
+          },
+          error: (error) => {
+            this.toastService.showTimedAlert('Xóa thất bại', error, 'error', 2000);
+          }
+        });
+        if(this.orderByUser){
+          this.cartDetailService.getUserOrderDetail(this.orderByUser?.orderId);
         }
-      },
-      error: (error) => {
-        this.toastService.showTimedAlert('Xóa thất bại', error, 'error', 2000);
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        
       }
     });
-    if(this.orderByUser){
-      this.cartDetailService.getUserOrder(this.orderByUser?.orderId);
-    }
+    
   }
 
   plusButtonHandler() {
@@ -201,20 +231,9 @@ export class UserCartComponent implements OnInit {
     });
   }
 
-  ngOnInit(): void {
-    this.setCartInfomationArrays();
-    this.cartService.getProvince();
-    this.cartService.provinceData$.subscribe(data => {
-      this.Provinces = Object.values(data);
-      console.log(this.Provinces);
-    });
-    this.cartService.districtData$.subscribe(data => {
-      this.Districts = Object.values(data);
-    });
-    this.cartService.wardData$.subscribe(data => {
-      this.Wards = Object.values(data);
-    });
+  getDiscount(discountId: number){
     
+    this.discountService.getDiscountById(discountId);
   }
 
   onProvinceChange() {
